requirements.txt
flask
waitress
black
psycopg2-binary
prometheus_client
alembic
pycryptodome
pyjwt
cryptography
marshmallow-sqlalchemy
